organizado pacote core da api
