+++ jacaEventos/urls.py
@@ -22,6 +22,6 @@ from django.contrib.auth import views as auth_views
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url(r'', include('user.urls'), name='usuario'),
-    url(r'^api/token/', obtain_auth_token, name='api-token'),
-    url(r'^api/', include(router.urls)),
+    url(r'^api/v1/token/', obtain_auth_token, name='api-token'),
+    url(r'^api/v1/', include(router.urls)),
 ]
